Semana7 aula26

+++ semana7/labenusers/src/components/UsersListPage.js
@@ -1,41 +1,46 @@
 import React from 'react'
 import axios from 'axios'
 import DeleteButton from './styled'
+import UserDetailsPage from './UserDetailsPage'
 
 const baseUrl = "https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users";
 const headersConfig =  {headers: {Authorization: "gregory-mendes-jackson"}};
 
 class UsersListPage extends React.Component {
     state = {
-        users: []
-    } 
+        users: [],
+    };
 
     componentDidMount = () => {
         this.getAllUsers();
-    }
+    };
     
     getAllUsers = () => {
-        const request = axios.get(baseUrl, headersConfig)
-    
+        const request = axios.get(baseUrl, headersConfig);
+        
         request
         .then(response => {
-          this.setState({users: response.data})
+            this.setState({users: response.data})
         })
         .catch(error => {
           alert("Ocorreu um erro inesperado.")
-        })
-    }
+        });
+    };
 
     deleteUser = (userId) => {
-        const request = axios.delete(`${baseUrl}/${userId}`, headersConfig)
-
-        request
-        .then(response => {
-            this.getAllUsers()
-        })
-        .catch(error => {
-            console.log(error.message)
-        })
+        const confirmDeletion = window.confirm("Tem certeza que deseja excluir este(a) usuário(a)?");
+        
+        if (confirmDeletion) {
+            const request = axios.delete(`${baseUrl}/${userId}`, headersConfig);
+            
+            request
+            .then(response => {
+                this.getAllUsers()
+            })
+            .catch(error => {
+                alert(error.message)
+            });
+        }
     }
 
     render() {
@@ -43,7 +48,7 @@ class UsersListPage extends React.Component {
             <div>
                 {this.state.users.map((user) => {
                     return (
-                        <p key={user.id}>
+                        <p key={user.id} onClick={this.onClickUserName}>
                             {user.name}
                             <DeleteButton onClick={() => this.deleteUser(user.id)}> X</DeleteButton>
                         </p>
